+++ src/coupon/application/use-cases/tier-1-in-domain/issue-user-coupon.use-case.spec.ts
@@ -15,7 +15,10 @@ import {
   UserCouponStatus,
 } from "@/coupon/domain/entities/user-coupon.entity";
 import { v4 as uuidv4 } from "uuid";
-import { IssueUserCouponDomainService } from "@/coupon/domain/services/issue-user-coupon.service";
+
+jest.mock("typeorm-transactional", () => ({
+  Transactional: () => () => ({}),
+}));
 
 describe("IssueUserCouponUseCase", () => {
   let useCase: IssueUserCouponUseCase;
@@ -35,7 +38,6 @@ describe("IssueUserCouponUseCase", () => {
     const module = await Test.createTestingModule({
       providers: [
         IssueUserCouponUseCase,
-        IssueUserCouponDomainService,
         {
           provide: "CouponRepositoryInterface",
           useValue: couponRepository,
