+++ src/coupon/coupon.module.ts
@@ -15,10 +15,7 @@ import { CouponTypeOrmEntity } from "./infrastructure/persistence/orm/coupon.typ
 import { UserCouponTypeOrmEntity } from "./infrastructure/persistence/orm/user-coupon.typeorm.entity";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { RecoverUserCouponUseCase } from "./application/use-cases/tier-1-in-domain/recover-user-coupon.use-case";
-import { UseUserCouponDomainService } from "./domain/services/use-user-coupon.service";
-import { ValidateUserCouponDomainService } from "./domain/services/validate-user-coupon.service";
-import { CancelUserCouponDomainService } from "./domain/services/cancel-user-coupon.service";
-import { IssueUserCouponDomainService } from "./domain/services/issue-user-coupon.service";
+import { ValidateUserCouponService } from "./domain/services/validate-user-coupon.service";
 
 @Module({
   imports: [
@@ -35,10 +32,7 @@ import { IssueUserCouponDomainService } from "./domain/services/issue-user-coupo
     ValidateCouponUseCase,
     CancelUserCouponUseCase,
     RecoverUserCouponUseCase,
-    CancelUserCouponDomainService,
-    UseUserCouponDomainService,
-    ValidateUserCouponDomainService,
-    IssueUserCouponDomainService,
+    ValidateUserCouponService,
     {
       provide: "UserCouponRepositoryInterface",
       useClass: UserCouponRepository,
