+++ src/coupon/domain/services/validate-user-coupon.service.ts
@@ -1,8 +1,8 @@
 import { Coupon } from "../entities/coupon.entity";
 import { UserCoupon } from "../entities/user-coupon.entity";
 
-export class ValidateUserCouponDomainService {
-  async validateUserCoupon({
+export class ValidateUserCouponService {
+  validateUserCoupon({
     coupon,
     userCoupon,
     orderPrice,
@@ -10,33 +10,10 @@ export class ValidateUserCouponDomainService {
     coupon: Coupon;
     userCoupon: UserCoupon;
     orderPrice: number;
-  }): Promise<{
-    isValid: boolean;
-    discountPrice: number;
-    discountedPrice: number;
-  }> {
-    if (!coupon.canUse(orderPrice)) {
-      return {
-        isValid: false,
-        discountPrice: 0,
-        discountedPrice: orderPrice,
-      };
+  }): boolean {
+    if (!coupon.canUse(orderPrice) || !userCoupon.canUse()) {
+      return false;
     }
-
-    if (!userCoupon.canUse()) {
-      return {
-        isValid: false,
-        discountPrice: 0,
-        discountedPrice: orderPrice,
-      };
-    }
-
-    const { discountPrice, discountedPrice } = coupon.use(orderPrice);
-
-    return {
-      isValid: true,
-      discountPrice,
-      discountedPrice,
-    };
+    return true;
   }
 }
